Update earned points calculation to float division
Previous version:
2 separate orders with costs 4.5 and 5.5 would not add points. New code accomodates decimal points without changing what user can see

+++ proj2/backend/app/main.py
@@ -473,7 +473,7 @@ def complete_run(
     # Calculate total bill and points
     orders = session.exec(select(Order).where(Order.run_id == run_id)).all()
     total_amount = sum(order.amount for order in orders)
-    earned_points = int(total_amount / 10)  # 1 point per $10
+    earned_points = total_amount / 10  # 1 point per $10
     
     # Update run status
     food_run.status = "completed"
